struct(T6): move event outcomes into events_apply.rs
Relocate apply_random_event and apply_historical_event out of turn.rs
into a new src/game/events_apply.rs (pub(super) impl Game, bodies
unchanged). Trigger selection stays in events.rs; week orchestration and
the two call sites stay in turn.rs (down from 538 to 285 lines). The
`mod events_apply;` wiring already landed with T1.

Extraction authored by pier-t6; reviewed here and verified green
(fmt/clippy/test) as part of landing T1+T6 together.
